LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/blob.https-expected.txt

CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/common/blank.html due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/common/blank.html due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/common/blank.html due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://localhost:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://localhost:9443/common/blank.html due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://localhost:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://localhost:9443/common/blank.html due to access control checks.
CONSOLE MESSAGE: Cancelled load to blob:https://127.0.0.1:9443/5b2efa55-73ac-4027-96d6-2a067e2b2c9e because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Cancelled load to blob:https://127.0.0.1:9443/f547926e-e784-402f-90bf-b888bb70401c because it violates the resource's Cross-Origin-Resource-Policy response header.

Harness Error (TIMEOUT), message = null

PASS Cross-Origin-Embedder-Policy and blob: URL from https://localhost:9443 in subframe via subframe
PASS Cross-Origin-Embedder-Policy and blob: URL from https://localhost:9443 in subframe via navigate
PASS Cross-Origin-Embedder-Policy and blob: URL from https://localhost:9443 in subframe via popup
FAIL Cross-Origin-Embedder-Policy and blob: URL from https://127.0.0.1:9443 in subframe via subframe assert_true: Cross-origin without CORP did not fail expected true got false
TIMEOUT Cross-Origin-Embedder-Policy and blob: URL from https://127.0.0.1:9443 in subframe via navigate Test timed out
FAIL Cross-Origin-Embedder-Policy and blob: URL from https://127.0.0.1:9443 in subframe via popup assert_true: Cross-origin without CORP did not fail expected true got false
PASS Cross-Origin-Embedder-Policy and blob: URL from https://127.0.0.1:9443 in subframe via subframe
TIMEOUT Cross-Origin-Embedder-Policy and blob: URL from https://127.0.0.1:9443 in subframe via navigate Test timed out
PASS Cross-Origin-Embedder-Policy and blob: URL from https://127.0.0.1:9443 in subframe via popup
